move mi test from test_ix.py
